Move common to node-services
node-services protocol is meant for internal restate communication. pb/ appears to go beyond the internal cluster scope.
