+++ crates/node-services/Cargo.toml
@@ -11,7 +11,6 @@ publish = false
 default = []
 
 [dependencies]
-restate-pb = { workspace = true, features = ["common"] }
 restate-types = { workspace = true }
 
 anyhow = { workspace = true, optional = true }
