+++ crates/node-services/proto/cluster_ctrl_svc.proto
@@ -9,7 +9,7 @@
 
 syntax = "proto3";
 
-import "dev/restate/common/common.proto";
+import "common.proto";
 
 package dev.restate.cluster_ctrl;
