+++ crates/node-services/src/lib.rs
@@ -8,13 +8,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use restate_pb::restate::common;
-
 pub mod cluster_ctrl {
-    #![allow(warnings)]
-    #![allow(clippy::all)]
-    #![allow(unknown_lints)]
-
     tonic::include_proto!("dev.restate.cluster_ctrl");
 
     pub const FILE_DESCRIPTOR_SET: &[u8] =
@@ -22,12 +16,53 @@ pub mod cluster_ctrl {
 }
 
 pub mod node {
-    #![allow(warnings)]
-    #![allow(clippy::all)]
-    #![allow(unknown_lints)]
-
     tonic::include_proto!("dev.restate.node");
 
     pub const FILE_DESCRIPTOR_SET: &[u8] =
         tonic::include_file_descriptor_set!("node_svc_descriptor");
+}
+
+pub mod common {
+    tonic::include_proto!("dev.restate.common");
+
+    pub const FILE_DESCRIPTOR_SET: &[u8] = tonic::include_file_descriptor_set!("common_descriptor");
+
+    impl From<Version> for restate_types::Version {
+        fn from(version: Version) -> Self {
+            restate_types::Version::from(version.value)
+        }
+    }
+
+    impl From<restate_types::Version> for Version {
+        fn from(version: restate_types::Version) -> Self {
+            Version {
+                value: version.into(),
+            }
+        }
+    }
+
+    impl From<NodeId> for restate_types::NodeId {
+        fn from(node_id: NodeId) -> Self {
+            restate_types::NodeId::new(node_id.id, node_id.generation)
+        }
+    }
+
+    impl From<restate_types::NodeId> for NodeId {
+        fn from(node_id: restate_types::NodeId) -> Self {
+            NodeId {
+                id: node_id.id().into(),
+                generation: node_id.as_generational().map(|g| g.generation()),
+            }
+        }
+    }
+
+    impl From<restate_types::PlainNodeId> for NodeId {
+        fn from(node_id: restate_types::PlainNodeId) -> Self {
+            let id: u32 = node_id.into();
+            NodeId {
+                id,
+                generation: None,
+            }
+        }
+    }
 }
