+++ crates/pb/build.rs
@@ -348,7 +348,6 @@ fn main() -> std::io::Result<()> {
                 "proto/dev/restate/internal/services.proto",
                 "proto/dev/restate/events.proto",
                 "proto/dev/restate/internal/messages.proto",
-                "proto/dev/restate/common/common.proto",
             ],
             &["proto"],
         )?;
