+++ crates/pb/src/lib.rs
@@ -70,54 +70,6 @@ pub mod restate {
             }
         }
     }
-
-    #[cfg(feature = "common")]
-    pub mod common {
-        #![allow(warnings)]
-        #![allow(clippy::all)]
-        #![allow(unknown_lints)]
-
-        include!(concat!(env!("OUT_DIR"), "/dev.restate.common.rs"));
-
-        impl From<Version> for restate_types::Version {
-            fn from(version: Version) -> Self {
-                restate_types::Version::from(version.value)
-            }
-        }
-
-        impl From<restate_types::Version> for Version {
-            fn from(version: restate_types::Version) -> Self {
-                Version {
-                    value: version.into(),
-                }
-            }
-        }
-
-        impl From<NodeId> for restate_types::NodeId {
-            fn from(node_id: NodeId) -> Self {
-                restate_types::NodeId::new(node_id.id, node_id.generation)
-            }
-        }
-
-        impl From<restate_types::NodeId> for NodeId {
-            fn from(node_id: restate_types::NodeId) -> Self {
-                NodeId {
-                    id: node_id.id().into(),
-                    generation: node_id.as_generational().map(|g| g.generation()),
-                }
-            }
-        }
-
-        impl From<restate_types::PlainNodeId> for NodeId {
-            fn from(node_id: restate_types::PlainNodeId) -> Self {
-                let id: u32 = node_id.into();
-                NodeId {
-                    id,
-                    generation: None,
-                }
-            }
-        }
-    }
 }
 
 pub static DESCRIPTOR_POOL: Lazy<DescriptorPool> = Lazy::new(|| {
